+++ t/634-bite-email-postfix.t
@@ -51,6 +51,7 @@ my $isexpected = [
     { 'n' => '53', 's' => qr/\A5[.]0[.]0\z/,    'r' => qr/syntaxerror/,'b' => qr/\A1\z/ },
     { 'n' => '54', 's' => qr/\A5[.]7[.]1\z/,    'r' => qr/rejected/,   'b' => qr/\A1\z/ },
     { 'n' => '55', 's' => qr/\A5[.]0[.]0\z/,    'r' => qr/toomanyconn/,'b' => qr/\A1\z/ },
+    { 'n' => '56', 's' => qr/\A4[.]4[.]2\z/,    'r' => qr/networkerror/,'b' => qr/\A1\z/ },
 ];
 
 $enginetest->($enginename, $isexpected);