+++ xt/634-bite-email-postfix.t
@@ -214,6 +214,9 @@ my $isexpected = [
     { 'n' => '01204', 'r' => qr/syntaxerror/    },
     { 'n' => '01205', 'r' => qr/rejected/       },
     { 'n' => '01206', 'r' => qr/toomanyconn/    },
+    { 'n' => '01207', 'r' => qr/toomanyconn/    },
+    { 'n' => '01208', 'r' => qr/toomanyconn/    },
+    { 'n' => '01209', 'r' => qr/networkerror/   },
 ];
 
 plan 'skip_all', sprintf("%s not found", $samplepath) unless -d $samplepath;